feat(sequence): support globalFilters array for main sequence
An array of globalFilters can be defined that are returned
as the first filters in getAllFilters().
The intended use-case are negative filters that limit the search
e.g. to only one ecu.

Sub-sequences can not contain globalFilters as this would
require a more sophisticated logic to query msgs for
sub-sequences vs the other sequence steps.